+++ packages/benches/p50/ai/p50-conversation-operations.bench.ts
@@ -91,7 +91,10 @@ bench('p50 conversation operations', async () => {
 
   await client.conversations.ChatBot.list();
 
-  conversation?.onMessage(() => { });
+  conversation?.onStreamEvent({
+    next: () => {},
+    error: () => {},
+  });
 
   await conversation?.sendMessage({
     content: [{ text: 'foo' }],