+++ packages/data-schema/src/ai/ConversationType.ts
@@ -13,6 +13,7 @@ import {
   ConversationSendMessageInputContent,
 } from './types/ConversationMessageContent';
 import { ToolConfiguration } from './types/ToolConfiguration';
+import { ConversationStreamErrorEvent, ConversationStreamEvent } from './types/ConversationStreamEvent';
 
 export const brandName = 'conversationCustomOperation';
 
@@ -23,6 +24,7 @@ export interface ConversationMessage {
   createdAt: string;
   id: string;
   role: 'user' | 'assistant';
+  associatedUserMessageId?: string;
 }
 
 // conversation route types
@@ -105,7 +107,10 @@ interface ConversationListMessagesInput {
   nextToken?: string | null;
 }
 
-type ConversationOnMessageHandler = (message: ConversationMessage) => void;
+type ConversationOnStreamEventHandler = {
+  next: (event: ConversationStreamEvent) => void;
+  error: (error: ConversationStreamErrorEvent) => void;
+};
 
 export interface Conversation {
   id: string;
@@ -133,9 +138,9 @@ export interface Conversation {
   /**
    * @experimental
    *
-   * Subscribes to new messages on the current conversation.
+   * Subscribes to new stream events on the current conversation.
    */
-  onMessage: (handler: ConversationOnMessageHandler) => Subscription;
+  onStreamEvent: (handler: ConversationOnStreamEventHandler) => Subscription;
 }
 
 // schema definition input
@@ -213,3 +218,5 @@ function _conversation(input: ConversationInput): ConversationType {
 export function conversation(input: ConversationInput): ConversationType {
   return _conversation(input);
 }
+export { ConversationStreamEvent };
+