+++ packages/data-schema/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aws-amplify/data-schema
 
+## 1.13.0
+
+### Minor Changes
+
+- 20f30fe: add streaming support for conversation routes
+- 20f30fe: propagate conversation errors through subscription
+
 ## 1.12.2
 
 ### Patch Changes